+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Remove `Box` usage from `State::coord_fns`.
 - Add `DDEMutStateFnMut`, which allows state-mutating events which have (non-mutating) access to history.
 - Add `d_prev` method for a `State` for free evaluation of derivative of the state at the beginning of the last computed step.
+- Rename `Equation::with_delay` method into `Equation::max_delay`
+- Add `Equation::const_delay` and `Equation::delay` methods
 
 # Version 0.0.2